+++ tests/testthat/test-intraday.R
@@ -1,4 +1,4 @@
-test_that("news", {
+test_that("intraday", {
 
   # test relies on calling api (we skip it on cran to save network bandwidth)
   skip_if_offline()
@@ -7,20 +7,20 @@ test_that("news", {
   suppressMessages({
     set_token()
 
-    df_news1 <- get_news(
+    df_intraday1 <- get_intraday(
       ticker = "AAPL",
       exchange = "US"
     )
 
-    expect_true(nrow(df_news1) > 0)
+    expect_true(nrow(df_intraday1) > 0)
 
     # run it again for testing local cache
-    df_news2 <- get_news(
+    df_intraday2 <- get_intraday(
       ticker = "AAPL",
       exchange = "US"
     )
 
-    expect_true(identical(df_news1, df_news2))
+    expect_true(identical(df_intraday1, df_intraday2))
 
 
   })