+++ NEWS.md
@@ -1,3 +1,7 @@
+## Version 0.5.2 (2025-07-28)
+
+- implemented `get_intraday()`, a function to download intraday prices from eodhd
+
 ## Version 0.5.1 (2024-09-03)
 
 - implemented `get_news()`